# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task API",
        "description": "API para gestionar tareas.",
        "version": "1.0.0"
    },
    "paths": {
        "/clients": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Lista todos los clientes",
                "operationId": "af1f3794367a5e843bf15e3a9033d911",
                "responses": {
                    "200": {
                        "description": "Lista de clientes.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Client"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Clients"
                ],
                "summary": "Almacena un nuevo cliente",
                "operationId": "15981a5da120d8b4ea1b91b05a35622b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreClientRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redirecciona a la lista de clientes con mensaje de éxito."
                    }
                }
            }
        },
        "/clients/{id}": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Muestra los detalles de un cliente",
                "operationId": "2965e9737f628d15ea88209095622c24",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del cliente.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del cliente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Client"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Clients"
                ],
                "summary": "Elimina un cliente",
                "operationId": "20de5e8bdef993807384fd7040f9ae92",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del cliente a eliminar.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirecciona a la lista de clientes."
                    }
                }
            }
        },
...
        "/clients/send-invoice/{id}": {
            "post": {
                "tags": [
                    "Clients"
                ],
                "summary": "Envía la factura de una cuota por correo electrónico",
                "operationId": "be4cf21741abc04b98483a4698fc8d1a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la cuota.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Factura enviada correctamente."
                    }
                }
            }
        },
        "/clients/{client}/edit": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Muestra el formulario para editar un cliente",
                "operationId": "27161bb9e8a4e561dfd01094a82c4d62",
                "parameters": [
                    {
                        "name": "client",
                        "in": "path",
                        "description": "ID del cliente a editar.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Formulario de edición de cliente."
                    }
                }
            }
        },
        "/clients/{client}": {
            "put": {
                "tags": [
                    "Clients"
                ],
                "summary": "Actualiza un cliente existente",
                "operationId": "558832db987fc02c8a5107fb013891c9",
                "parameters": [
                    {
                        "name": "client",
                        "in": "path",
                        "description": "ID del cliente a actualizar.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ModifyClientRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redirecciona a la lista de clientes con mensaje de éxito."
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ModifyClientRequest": {
                "title": "Modify Client Request",
                "description": "Request para modificar un cliente existente.",
                "required": [
                    "cif",
                    "name",
                    "phone",
                    "email",
                    "monthly_fee"
                ],
                "properties": {
                    "cif": {
                        "description": "CIF del cliente.",
                        "type": "string"
                    },
                    "name": {
                        "description": "Nombre del cliente.",
                        "type": "string"
                    },
                    "phone": {
                        "description": "Teléfono del cliente (9 dígitos).",
                        "type": "string"
                    },
                    "email": {
                        "description": "Correo electrónico del cliente.",
                        "type": "string",
                        "format": "email"
                    },
                    "bank_account": {
                        "description": "Cuenta bancaria del cliente (opcional).",
                        "type": "string"
                    },
                    "country": {
                        "description": "País del cliente (opcional).",
                        "type": "string"
                    },
                    "currency": {
                        "description": "Moneda del cliente (opcional).",
                        "type": "string"
                    },
                    "monthly_fee": {
                        "description": "Cuota mensual del cliente.",
                        "type": "number",
                        "format": "float"
                    }
                },
                "type": "object"
            },
...
            "StoreClientRequest": {
                "title": "Store Client Request",
                "description": "Request para almacenar un nuevo cliente.",
                "required": [
                    "cif",
                    "name",
                    "phone",
                    "email",
                    "bank_account",
                    "country",
                    "monthly_fee"
                ],
                "properties": {
                    "cif": {
                        "description": "CIF del cliente.",
                        "type": "string"
                    },
                    "name": {
                        "description": "Nombre del cliente.",
                        "type": "string"
                    },
                    "phone": {
                        "description": "Teléfono del cliente (9 dígitos).",
                        "type": "string"
                    },
                    "email": {
                        "description": "Correo electrónico del cliente.",
                        "type": "string",
                        "format": "email"
                    },
                    "bank_account": {
                        "description": "Cuenta bancaria del cliente.",
                        "type": "string"
                    },
                    "country": {
                        "description": "País del cliente.",
                        "type": "string"
                    },
                    "currency": {
                        "description": "Moneda del cliente (opcional).",
                        "type": "string"
                    },
                    "monthly_fee": {
                        "description": "Cuota mensual del cliente.",
                        "type": "number",
                        "format": "float"
                    }
                },
                "type": "object"
            },
...
            "Client": {
                "title": "Client",
                "description": "Modelo de cliente.",
                "properties": {
                    "id": {
                        "description": "ID del cliente.",
                        "type": "integer"
                    },
                    "cif": {
                        "description": "CIF del cliente.",
                        "type": "string"
                    },
                    "name": {
                        "description": "Nombre del cliente.",
                        "type": "string"
                    },
                    "phone": {
                        "description": "Teléfono del cliente.",
                        "type": "string"
                    },
                    "email": {
                        "description": "Correo electrónico del cliente.",
                        "type": "string",
                        "format": "email"
                    },
                    "bank_account": {
                        "description": "Cuenta bancaria del cliente.",
                        "type": "string"
                    },
                    "country": {
                        "description": "País del cliente.",
                        "type": "string"
                    },
                    "currency": {
                        "description": "Moneda del cliente.",
                        "type": "string"
                    },
                    "monthly_fee": {
                        "description": "Cuota mensual del cliente.",
                        "type": "number",
                        "format": "decimal"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}